Added call to postSurvey for assigned tasks
